+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4 (2015-09-02)
+
+##### New
+- Add method -selectedRowAtColumn:
+
+
 ## 0.0.3 (2015-09-02)
 
 ##### New